refactor workbench sandbox changes router
